Add files via upload
I added in slides for the code I had prepared for the workshop. Some of it (apply functions) looks redundant with the stuff you put together. I had written down that I was supposed to be covering loops, functionals, and logicals, but I can certainly take it out if you think it overlaps with what you have to much. Although I did not think I had time for the purrr functions, so maybe it is okay. I am not a big fan of outlines at the start of a lecture, but if you want them for your sake I am okay with them. There were several spelling errors in what you have. I did not fix them, but if you would like me to go through and proofread let me know and I can. I would like to test the code I put together within Quarto once you approve the commit (if you are okay with what I changed). I know it works in webR because tested it a couple of weeks ago, but I changed some stuff to make it work in code chunks that I would like to test. Lastly, I did not add a table of the infix operators. I think that most people attending an "intro to advanced R" should have the basics down and most people are not going to memorize a table when it is so easy to look them up in the documentation. I can switch out the functionals for how to create user defined infix operators (although that might make more sense after the functions have been covered).